# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Football API - AYO Technical Test 2026",
    "description": "Backend API untuk manajemen tim sepakbola amatir perusahaan XYZ. Menyediakan CRUD untuk Tim, Pemain, Pertandingan, pencatatan Gol, dan Laporan hasil pertandingan.\n\nSemua endpoint GET bersifat publik. Endpoint POST/PUT/DELETE memerlukan header `X-API-Key` (lihat Authorize di kanan atas).",
    "version": "1.0",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header",
      "description": "API Key untuk endpoint yang membutuhkan otorisasi (create/update/delete)."
    }
  },
...
  "paths": {
...
    "/matches": {
      "get": {
        "summary": "Daftar semua pertandingan",
        "tags": [
          "Matches"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MatchListResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Buat jadwal pertandingan",
        "tags": [
          "Matches"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateMatchRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/MatchResponse"
            }
          },
          "400": {
            "description": "Bad Request - tim home dan away sama",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found - tim tidak ditemukan",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/matches/{id}": {
      "get": {
        "summary": "Detail pertandingan",
        "tags": [
          "Matches"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID pertandingan"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MatchResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Update jadwal pertandingan",
        "tags": [
          "Matches"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID pertandingan"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UpdateMatchRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MatchResponse"
            }
          },
          "400": {
            "description": "Bad Request - tim home dan away sama",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity - pertandingan sudah completed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Hapus jadwal pertandingan (soft delete)",
        "tags": [
          "Matches"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID pertandingan"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeletedResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/matches/{id}/result": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Submit hasil pertandingan",
        "description": "Menyimpan skor akhir dan daftar pencetak gol. Jumlah gol pemain dari tim home harus sama dengan home_score, begitu juga away_score untuk tim away. Pertandingan otomatis berubah status menjadi `completed` setelah hasil disimpan.",
        "tags": [
          "Matches"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID pertandingan"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SubmitResultRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GoalListResponse"
            }
          },
          "400": {
            "description": "Bad Request - pemain bukan dari tim yang bertanding / jumlah gol tidak sesuai skor",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found - pertandingan tidak ditemukan",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity - pertandingan sudah selesai",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Team": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Persib Bandung"
        },
        "logo": {
          "type": "string",
          "example": "https://example.com/logo.png"
        },
        "founded": {
          "type": "integer",
          "example": 1933
        },
        "headquarters_address": {
          "type": "string",
          "example": "Jl. Sulanjana No.17"
        },
        "headquarters_city": {
          "type": "string",
          "example": "Bandung"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Player": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 3
        },
        "team_id": {
          "type": "integer",
          "example": 1
        },
        "team": {
          "$ref": "#/definitions/Team"
        },
        "name": {
          "type": "string",
          "example": "Achmad Jufriyanto"
        },
        "height": {
          "type": "number",
          "example": 178
        },
        "weight": {
          "type": "number",
          "example": 72
        },
        "position": {
          "type": "string",
          "enum": [
            "penyerang",
            "gelandang",
            "bertahan",
            "penjaga_gawang"
          ],
          "example": "bertahan"
        },
        "jersey_number": {
          "type": "integer",
          "example": 5
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Goal": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "match_id": {
          "type": "integer",
          "example": 1
        },
        "player_id": {
          "type": "integer",
          "example": 3
        },
        "team_id": {
          "type": "integer",
          "example": 1
        },
        "minute": {
          "type": "integer",
          "example": 23
        },
        "player": {
          "$ref": "#/definitions/Player"
        },
        "team": {
          "$ref": "#/definitions/Team"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Match": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "match_date": {
          "type": "string",
          "format": "date-time",
          "example": "2026-07-15T15:00:00Z"
        },
        "home_team_id": {
          "type": "integer",
          "example": 1
        },
        "away_team_id": {
          "type": "integer",
          "example": 2
        },
        "home_team": {
          "$ref": "#/definitions/Team"
        },
        "away_team": {
          "$ref": "#/definitions/Team"
        },
        "home_score": {
          "type": "integer",
          "example": 2
        },
        "away_score": {
          "type": "integer",
          "example": 1
        },
        "status": {
          "type": "string",
          "enum": [
            "scheduled",
            "completed"
          ],
          "example": "scheduled"
        },
        "goals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Goal"
          }
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "CreateMatchRequest": {
      "type": "object",
      "required": [
        "match_date",
        "home_team_id",
        "away_team_id"
      ],
      "properties": {
        "match_date": {
          "type": "string",
          "format": "date-time",
          "example": "2026-07-15T15:00:00Z"
        },
        "home_team_id": {
          "type": "integer",
          "example": 1
        },
        "away_team_id": {
          "type": "integer",
          "example": 2
        }
      }
    },
    "UpdateMatchRequest": {
      "type": "object",
      "properties": {
        "match_date": {
          "type": "string",
          "format": "date-time"
        },
        "home_team_id": {
          "type": "integer"
        },
        "away_team_id": {
          "type": "integer"
        }
      }
    },
    "GoalRequest": {
      "type": "object",
      "required": [
        "player_id",
        "minute"
      ],
      "properties": {
        "player_id": {
          "type": "integer",
          "example": 3
        },
        "minute": {
          "type": "integer",
          "minimum": 1,
          "maximum": 120,
          "example": 23
        }
      }
    },
    "SubmitResultRequest": {
      "type": "object",
      "required": [
        "home_score",
        "away_score",
        "goals"
      ],
      "properties": {
        "home_score": {
          "type": "integer",
          "minimum": 0,
          "example": 2
        },
        "away_score": {
          "type": "integer",
          "minimum": 0,
          "example": 1
        },
        "goals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GoalRequest"
          }
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "request failed"
        },
        "error": {
          "type": "string",
          "example": "terjadi kesalahan"
        }
      }
    },
...
    "MatchResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "OK"
        },
        "data": {
          "$ref": "#/definitions/Match"
        }
      }
    },
    "MatchListResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "OK"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Match"
          }
        }
      }
    },
    "GoalListResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "OK"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Goal"
          }
        }
      }
    },
...
    "DeletedResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "OK"
        }
      }
    }
  }
}